pandora: Update to use latest net mgmt API
A change has been done the net mgmt API.
This updates the service to be compatible with it.

+++ subsys/net/lib/hermes/wifi.c
@@ -38,7 +38,7 @@ static void hermes_wifi_handler_put_credentials(struct hermes_resource *rsc, con
 	pandora_wifi_set_credentials(&credentials);
 }
 
-static void wifi_event_handler(struct net_mgmt_event_callback *cb, uint32_t mgmt_event,
+static void wifi_event_handler(struct net_mgmt_event_callback *cb, uint64_t mgmt_event,
 			       struct net_if *iface)
 {
 	switch (mgmt_event) {
